separated task refiner from baseclass refiner file.
